+++ src/Model/Label/EventTypes.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace TK\GitHubWebhook\Model\Label;
+
+enum EventTypes: string
+{
+    case CREATED = "created";
+    case DELETED = "deleted";
+    case EDITED = "edited";
+}